editor/libeditor/tests/test_bug1620778.html

<!DOCTYPE html>
<title>Test for Bug 1620778</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="/tests/SimpleTest/EventUtils.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<input id=a value=abcd autocomplete=off>
<input id=a value=abcd>
<script>
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(() => {
  let expectedPosition = null;
  for (let input of document.querySelectorAll("input")) {
    input.focus();
    input.selectionStart = 0;
    synthesizeKey("KEY_ArrowRight");
    synthesizeKey("KEY_ArrowRight");
    synthesizeKey("KEY_ArrowDown");
    if (expectedPosition === null)
      expectedPosition = input.selectionStart;
    isnot(input.selectionStart, 0);
    is(input.selectionStart, expectedPosition, "autocomplete shouldn't make a difference on inputs that have no completion results of any kind");
  }
  SimpleTest.finish();
});
</script>
</body>
</html>
